# docs/swagger.yaml
basePath: /
definitions:
  models.ChangePasswordRequest:
    properties:
      email:
        type: string
      new_password:
        minLength: 6
        type: string
      old_password:
        minLength: 6
        type: string
    required:
    - email
    - new_password
    - old_password
    type: object
  models.CreateOrderRequest:
    properties:
      cinema_id:
        type: integer
      now_showing_id:
        type: integer
      payment_id:
        type: integer
      price:
        minimum: 0
        type: number
      seats_map:
        items:
          type: string
        minItems: 1
        type: array
      users_id:
        type: integer
    required:
    - now_showing_id
    - payment_id
    - price
    - seats_map
    type: object
  models.CreateOrderResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      price:
        type: number
      qr_code:
        type: string
      seats_map:
        items:
          type: string
        type: array
      ticket_id:
        type: integer
      users_id:
        type: integer
    type: object
...
  models.Profile:
    properties:
      created_at:
        example: "2025-09-14T12:00:00Z"
        type: string
      first_name:
        example: John
        type: string
      id:
        example: 1
        type: integer
      last_name:
        example: Doe
        type: string
      phone_number:
        example: "+628123456789"
        type: string
      profile_picture:
        example: profile_123.png
        type: string
      updated_at:
        example: "2025-09-14T12:30:00Z"
        type: string
    type: object
...
  models.UserProfileResponse:
    properties:
      email:
        example: john@example.com
        type: string
      first_name:
        example: John
        type: string
      id:
        example: 1
        type: integer
      last_name:
        example: Doe
        type: string
      phone_number:
        example: "+628123456789"
        type: string
      poin:
        example: 120
        type: integer
      profile_picture:
        example: profile_123.png
        type: string
      role:
        example: user
        type: string
    type: object
host: localhost:9001
info:
  contact: {}
  title: Backend Golang Tickitz App
  version: "1.0"
paths:
...
  /orders:
    post:
      consumes:
      - application/json
      description: Membuat order baru beserta ticket, update showing_seats, dan relasi
        ke cinema
      parameters:
      - description: Order Request
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/models.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Order berhasil dibuat
          schema:
            $ref: '#/definitions/models.CreateOrderResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create new order
      tags:
      - Orders
  /orders/history:
    get:
      consumes:
      - application/json
      description: Ambil riwayat pesanan berdasarkan user yang sedang login (dari
        JWT)
      produces:
      - application/json
      responses:
        "200":
          description: Success response with order history
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized, token tidak valid
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Tidak ada riwayat pesanan
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get user order history
      tags:
      - Orders
  /orders/seats/{now_showing_id}:
    get:
      consumes:
      - application/json
      description: Mengambil daftar kursi berdasarkan `now_showing_id`.
      parameters:
      - description: ID Now Showing
        in: path
        name: now_showing_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Mendapatkan semua kursi
      tags:
      - Seats
  /profile:
    get:
      description: Ambil data profil user dari JWT
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserProfileResponse'
      security:
      - BearerAuth: []
      summary: Get My Profile
      tags:
      - Profile
    patch:
      consumes:
      - multipart/form-data
      description: Update data profil user (dengan upload gambar opsional)
      parameters:
      - description: First Name
        in: formData
        name: first_name
        required: true
        type: string
      - description: Last Name
        in: formData
        name: last_name
        required: true
        type: string
      - description: Phone Number
        in: formData
        name: phone_number
        required: true
        type: string
      - description: Profile Picture
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Profile'
      security:
      - BearerAuth: []
      summary: Update User Profile
      tags:
      - Profile
  /profile/change-password:
    patch:
      consumes:
      - application/json
      description: Ubah password user
      parameters:
      - description: Password Data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password berhasil diubah
          schema:
            type: string
      summary: Change Password
      tags:
      - Profile
securityDefinitions:
  BearerAuth:
    description: 'Masukkan format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"